- implement '--load <PATH>' flag that loads an interaction plan and executes it instead of generating one from scratch - save a json serialization of the generated plans to `<tempdir>/simulator.plan.json`
